Separando configurações em arquivo separado
